requirements.txt
Django==2.2.27
djangorestframework==3.13.1
psycopg2-binary==2.8.6
django-environ==0.8.1
celery==5.2.3
redis==4.1.4
